Correct datatypes

+++ Homework2/src/transpose.cu
@@ -97,7 +97,7 @@ bool checkCorrectness(int* A, int* A_T, int size){
 
     for (int i=0; i<size; i++){
         for (int j=0; j<size; j++){
-            if (A_T[i * size + j] != res[i * size + j]) {
+            if (A_T[i * size + j] != (int) res[i * size + j]) {
                 return false;
             }
         }
